+++ Project_Blind/Assets/Scripts/Core/Sight/SightController.cs
@@ -1,3 +1,4 @@
+using System;
 using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
@@ -7,103 +8,37 @@ namespace Blind
 {
     public class SightController : SingletonDontCreate<SightController>
     {
-        public int resolution = 512;
-        [Range(0.01f, 1f)]
-        public float brushSize = 0.1f;
-        public bool ShowGizmos = true;
-        
-        public const float CurrentAlpha = 0f; // 현재 위치한 경우 알파값
-        public const float NeverAlpha = 1f; // 방문한 적 없는 경우 알파값
-
-        public LayerMask _groundLayer;
-
+         public bool ShowGizmos = true;
         [Space(8f)]
         public float TileSize = 1; // 타일 하나의 크기
-        
-        public float Offset = 3f; // 타일 하나의 크기
 
-        
-        private Texture2D _brushTexture;
-        private Texture2D _mainTex;
-        private Texture2D _originalRT;
-        private MeshRenderer _mr;
-        private RenderTexture _rt;
-        private RenderTexture _preRt;
-        
-        private Vector3 _size;
-        private Vector3 _center;
-        private Vector3 _origin;
-        private Bounds _bounds;
-        
         private List<TilePos> _visibleTiles = new List<TilePos>();
         private List<SightUnit> _unitList = new List<SightUnit>();
 
+        private Vector3 _size;
+        private Vector3 _origin;
+        private Vector3 _center;
+        private Bounds _bounds;
+
         private FOWMap _fowMap;
+        private FOWRenderer _firstLayer;
 
         protected override void Awake()
         {
             base.Awake();
-            GetComponent<MeshRenderer>().enabled = true;
-            TryGetComponent(out _mr);
+            _firstLayer = transform.GetChild(0).GetComponent<FOWRenderer>();
+        }
 
-            _rt = new RenderTexture(resolution, resolution, 32);
-            _preRt  = new RenderTexture(resolution,resolution,32);
-
-            if (_mr.material.mainTexture != null)
-            {
-                _mainTex = _mr.material.mainTexture as Texture2D;
-            }
-            // 메인 텍스쳐가 없을 경우, 하얀 텍스쳐를 생성하여 사용
-            else
-            {
-                _mainTex = new Texture2D(resolution, resolution);
-            }
-
-            _bounds = _mr.bounds;
-            _originalRT = new Texture2D(_mainTex.width, _mainTex.height);
-            
-            Color32[] resetColorArray = _originalRT.GetPixels32();
-
-            for (int i = 0; i < resetColorArray.Length; i++) {
-                resetColorArray[i] = Color.white;
-            }
-      
-            _originalRT.SetPixels32(resetColorArray);
-            _originalRT.Apply();
-
-            Graphics.Blit(_originalRT,_preRt);
-            Graphics.Blit(_originalRT,_rt);
-
-            _mr.material.SetTexture("_RenderTexture",_rt);
-
-            // 시야 텍스쳐 추가
-            _brushTexture = new Texture2D(resolution, resolution);
-            int rSquared = (resolution/2) * (resolution/2);
-            int x = resolution / 2, y = x;
-            for (int u = 0; u < resolution; u++)
-            {
-                for (int v = 0; v < resolution; v++)
-                {
-                    if ((x-u)*(x-u) + (y-v)*(y-v) < rSquared) {
-                        _brushTexture.SetPixel(u, v, new Color(0,0,0,1f));
-                    }
-                    else
-                    {
-                        _brushTexture.SetPixel(u, v, new Color(0,0,0,0));
-                    }
-                }
-            }
-            _brushTexture.Apply();
-            
-            // size측정
-            var bounds = _mr.bounds;
-            _size = bounds.size;
-            _center = bounds.center;
+        private void Start()
+        {
+            _size = _firstLayer.Size;
+            _bounds = _firstLayer.Bounds;
+            _center = _bounds.center;
             int sizeX = (int)(_size.x / (float)TileSize);
             int sizeY = (int)(_size.y / (float)TileSize);
             _fowMap = new FOWMap(sizeX,sizeY);
             _origin = _center - (Vector3)_fowMap.MapSize / 2;
-
+            
             FowMapInit();
             StartCoroutine(UpdateFog());
         }
@@ -113,25 +48,28 @@ namespace Blind
             while (true)
             {
                 _visibleTiles.Clear();
-                foreach(var unit in _unitList)
+                foreach (var unit in _unitList)
                 {
                     for (int i = -unit.Range; i < unit.Range; i++)
                     {
                         for (int j = -unit.Range; j < unit.Range; j++)
                         {
-                            _visibleTiles.Add(GetUnitTilePos(unit.transform) + new TilePos(i,j));
+                            _visibleTiles.Add(GetUnitTilePos(unit.transform) + new TilePos(i, j));
                         }
                     }
                 }
+
                 var positions = new List<Vector2>();
                 foreach (var tile in _visibleTiles)
                 {
                     positions.Add(GetTileCenterPos(tile));
                 }
-                DrawTexture(positions); yield return new WaitForSeconds(0.02f); 
+
+                _firstLayer.DrawTexture(positions);
+                yield return new WaitForSeconds(0.02f);
             }
         }
-        
+
         /// <summary>
         /// 맵 초기화
         /// </summary>
@@ -143,9 +81,9 @@ namespace Blind
                 for (var j = -1; j < mapSize.y; j++)
                 {
                     int layermask = 1 << LayerMask.NameToLayer("Floor");
-                    var ray = new Ray(GetTileCenterPos(i, j), new Vector3(0,0,1) * 10);
+                    var ray = new Ray(GetTileCenterPos(i, j), new Vector3(0, 0, 1) * 10);
                     RaycastHit hit;
-                    var raycast = Physics.Raycast(ray,out hit);
+                    var raycast = Physics.Raycast(ray, out hit);
                     if (raycast)
                     {
                         Debug.Log(10);
@@ -161,8 +99,8 @@ namespace Blind
 
         private TilePos GetUnitTilePos(Transform pos)
         {
-            var tmp = (pos.position - _origin)/TileSize;
-            return new TilePos((int)tmp.x, (int)tmp.y);
+            var tmp = (pos.position - _origin) / TileSize;
+            return new TilePos((int) tmp.x, (int) tmp.y);
         }
 
         private Vector2 GetTileCenterPos(TilePos tile)
@@ -170,88 +108,33 @@ namespace Blind
             return GetTileCenterPos(tile.x, tile.y);
         }
 
-        private Vector2 GetTileCenterPos(int x,int y)
+        private Vector2 GetTileCenterPos(int x, int y)
         {
-            var res = new Vector2(x * TileSize + TileSize/2 , y * TileSize + TileSize / 2);
-            res = (Vector2)_origin + res;
+            var res = new Vector2(x * TileSize + TileSize / 2, y * TileSize + TileSize / 2);
+            res = (Vector2) _origin + res;
             return res;
         }
 
         private void OnDrawGizmos()
         {
             if (Application.isPlaying == false) return;
-            
+
             foreach (var tile in _visibleTiles)
             {
                 Vector2 pos = GetTileCenterPos(tile.x, tile.y);
                 Gizmos.color = Color.green;
-                Gizmos.DrawCube(new Vector3(pos.x,pos.y,0f), new Vector3(TileSize, TileSize,1f));
+                Gizmos.DrawCube(new Vector3(pos.x, pos.y, 0f), new Vector3(TileSize, TileSize, 1f));
             }
-            
+
             if (ShowGizmos == false) return;
             for (int i = 0; i < _fowMap.MapSizeX; i++)
             {
                 for (int j = 0; j < _fowMap.MapSizeY; j++)
                 {
                     Gizmos.color = Color.white;
-                    Gizmos.DrawWireCube(GetTileCenterPos(i,j),new Vector3(TileSize,TileSize,0));
+                    Gizmos.DrawWireCube(GetTileCenterPos(i, j), new Vector3(TileSize, TileSize, 0));
                 }
             }
-        }
-
-        public void DrawTexture(IEnumerable<Vector2> positions)
-        {
-            var res = new List<Vector2>();
-            Profiler.BeginSample("UvPoint Calc");
-            foreach (var pos in positions)
-            {
-                Vector2 size = (_bounds.max - _bounds.min);
-                Vector2 uvPoint = ((Vector3)pos - _bounds.min) / size;
-                uvPoint = new Vector2(1,1)-uvPoint;
-                uvPoint *= resolution;
-                res.Add(uvPoint);
-            }
-            Profiler.EndSample();
-            _DrawTexture(res);
-        }
-
-
-        /// <summary> 렌더 텍스쳐에 브러시 텍스쳐로 그리기 </summary>
-        private void _DrawTexture(in IEnumerable<Vector2> uvList)
-        {
-            var sightSetAlpha = Resources.Load<Material>("Materials/Sight/SightSetAlpha");
-            sightSetAlpha.SetFloat("_Alpha",0.5f);
-            Material blurMat = Resources.Load<Material>("Materials/Sight/SightBlur");
-            blurMat.SetFloat("_Offset",Offset);
-            
-            Graphics.Blit(_preRt, _rt);
-            RenderTexture.active = _rt; // 페인팅을 위해 활성 렌더 텍스쳐 임시 할당
-            GL.PushMatrix();                                  // 매트릭스 백업
-            GL.LoadPixelMatrix(0, resolution, resolution, 0); // 알맞은 크기로 픽셀 매트릭스 설정
-
-            float brushPixelSize = brushSize * resolution;
-            
-            Profiler.BeginSample("_DrawTexture");
-            // 렌더 텍스쳐에 브러시 텍스쳐를 이용해 그리기
-            foreach (var uv in uvList)
-            {
-                Graphics.DrawTexture(
-                    new Rect(
-                        uv.x - brushPixelSize * 0.5f,
-                        (_rt.height - uv.y) - brushPixelSize * 0.5f,
-                        brushPixelSize,
-                        brushPixelSize
-                    ),
-                    _brushTexture
-                );
-            }
-            Profiler.EndSample();
-            GL.PopMatrix();              // 매트릭스 복구
-            RenderTexture.active = null; // 활성 렌더 텍스쳐 해제
-            
-            Graphics.Blit(_rt,_preRt,sightSetAlpha,0);
-            Graphics.Blit(_rt,_rt,blurMat,0);
-            Graphics.Blit(_rt,_rt,blurMat,0);
         }
     }
 }
