--- Project_Blind/Assets/Scripts/Core/Sight/test class.cs
@@ -1,7 +1,0 @@
-﻿namespace Blind.Sight
-{
-    public class test_class
-    {
-        
-    }
-}
